+++ CHANGELOG.md
@@ -179,6 +179,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Docker desktop from an NL intent; the desktop driver is a *connector*, the schema in
   the action space is the only core change).
 
+## [0.4.96] - 2026-06-24
+
+### Docs
+- Update README.md
+
+### Test
+- Update tests/test_host_dashboard.py
+
+### Other
+- Update .urirun/scheme-index.json
+- Update adapters/python/tests/test_exec.py
+- Update adapters/python/urirun/host/host_dashboard.py
+- Update adapters/python/urirun/runtime/v2.py
+
 ## [0.4.95] - 2026-06-24
 
 ### Docs